+++ packages/explorer/src/app/(explorer)/[chainName]/worlds/[worldAddress]/explore/TableSelector.tsx
@@ -1,7 +1,10 @@
-import { Check, ChevronsUpDown, Lock } from "lucide-react";
+import { CheckIcon, ChevronsUpDownIcon, Link2Icon, Link2OffIcon } from "lucide-react";
 import { useParams } from "next/navigation";
+import { useQueryState } from "nuqs";
+import { Hex } from "viem";
 import { useState } from "react";
-import { internalTableNames } from "@latticexyz/store-sync/sqlite";
+import { useEffect } from "react";
+import { Table } from "@latticexyz/config";
 import { Button } from "../../../../../../components/ui/Button";
 import {
   Command,
@@ -12,24 +15,54 @@ import {
   CommandList,
 } from "../../../../../../components/ui/Command";
 import { Popover, PopoverContent, PopoverTrigger } from "../../../../../../components/ui/Popover";
-import { cn } from "../../../../../../lib/utils";
+import { cn } from "../../../../../../utils";
+import { useChain } from "../../../../hooks/useChain";
+import { constructTableName } from "../../../../utils/constructTableName";
 
-type Props = {
-  value: string | undefined;
-  options: string[];
-};
+function TableSelectorItem({ table, selected, asOption }: { table: Table; selected: boolean; asOption?: boolean }) {
+  const { type, name, namespace } = table;
+  return (
+    <div className="flex items-center">
+      {asOption && <CheckIcon className={cn("mr-2 h-4 w-4", selected ? "opacity-100" : "opacity-0")} />}
+      {type === "offchainTable" && <Link2OffIcon className="mr-2 inline-block opacity-70" size={14} />}
+      {type === "table" && <Link2Icon className="mr-2 inline-block opacity-70" size={14} />}
+      {name} {namespace && <span className="ml-2 opacity-70">({namespace})</span>}
+    </div>
+  );
+}
 
-export function TableSelector({ value, options }: Props) {
-  const [open, setOpen] = useState(false);
+export function TableSelector({ tables }: { tables?: Table[] }) {
   const { worldAddress } = useParams();
+  const { id: chainId } = useChain();
+  const [selectedTableId, setTableId] = useQueryState("tableId");
+  const [open, setOpen] = useState(false);
+  const selectedTableConfig = tables?.find(({ tableId }) => tableId === selectedTableId);
+
+  useEffect(() => {
+    if (!selectedTableId && Array.isArray(tables) && tables.length > 0) {
+      setTableId(tables[0].tableId);
+    }
+  }, [selectedTableId, setTableId, tables]);
 
   return (
     <div className="w-full py-4">
       <Popover open={open} onOpenChange={setOpen}>
         <PopoverTrigger asChild>
-          <Button variant="outline" role="combobox" aria-expanded={open} className="w-full justify-between">
-            {value ? options.find((option) => option === value)?.replace(`${worldAddress}__`, "") : "Select table..."}
-            <ChevronsUpDown className="ml-2 h-4 w-4 shrink-0 opacity-50" />
+          <Button
+            variant="outline"
+            role="combobox"
+            aria-expanded={open}
+            className="w-full justify-between"
+            disabled={!tables}
+          >
+            {selectedTableConfig && (
+              <TableSelectorItem
+                table={selectedTableConfig}
+                selected={selectedTableId === selectedTableConfig.tableId}
+              />
+            )}
+            {!selectedTableConfig && <span className="opacity-50">Select table...</span>}
+            <ChevronsUpDownIcon className="ml-2 h-4 w-4 shrink-0 opacity-50" />
           </Button>
         </PopoverTrigger>
 
@@ -37,28 +70,20 @@ export function TableSelector({ value, options }: Props) {
           <Command>
             <CommandInput placeholder="Search tables..." className="font-mono" />
             <CommandList>
-              <CommandEmpty>No framework found.</CommandEmpty>
+              <CommandEmpty className="py-4 text-center font-mono text-sm">No table found.</CommandEmpty>
               <CommandGroup>
-                {options.map((option) => {
+                {tables?.map((table) => {
                   return (
                     <CommandItem
-                      key={option}
-                      value={option}
-                      onSelect={(currentValue) => {
-                        const url = new URL(window.location.href);
-                        const searchParams = new URLSearchParams(url.search);
-                        searchParams.set("tableId", currentValue);
-                        window.history.pushState({}, "", `${window.location.pathname}?${searchParams}`);
-
+                      key={table.tableId}
+                      value={constructTableName(table, worldAddress as Hex, chainId)}
+                      onSelect={() => {
+                        setTableId(table.tableId);
                         setOpen(false);
                       }}
                       className="font-mono"
                     >
-                      <Check className={cn("mr-2 h-4 w-4", value === option ? "opacity-100" : "opacity-0")} />
-                      {(internalTableNames as string[]).includes(option) && (
-                        <Lock className="mr-2 inline-block opacity-70" size={14} />
-                      )}
-                      {option.replace(`${worldAddress}__`, "")}
+                      <TableSelectorItem table={table} selected={selectedTableId === table.tableId} asOption />
                     </CommandItem>
                   );
                 })}