+++ packages/explorer/src/app/(explorer)/hooks/useChain.ts
@@ -1,6 +1,6 @@
 import { useParams } from "next/navigation";
 import { Chain } from "viem";
-import { supportedChains, validateChainName } from "../common";
+import { supportedChains, validateChainName } from "../../../common";
 
 export function useChain(): Chain {
   const { chainName } = useParams();