+++ packages/explorer/src/common.ts
@@ -1,4 +1,7 @@
-import { anvil, garnet, redstone } from "viem/chains";
+import { anvil } from "viem/chains";
+import { garnet, redstone } from "@latticexyz/common/chains";
+
+export const internalNamespaces = ["world", "store", "metadata"];
 
 export const supportedChains = { anvil, garnet, redstone } as const;
 export type supportedChains = typeof supportedChains;
@@ -8,7 +11,7 @@ export type supportedChainId = supportedChains[supportedChainName]["id"];
 
 export const chainIdToName = Object.fromEntries(
   Object.entries(supportedChains).map(([chainName, chain]) => [chain.id, chainName]),
-);
+) as Record<supportedChainId, supportedChainName>;
 
 export function validateChainId(chainId: unknown): asserts chainId is supportedChainId {
   if (!(typeof chainId === "number" && chainId in chainIdToName)) {