+++ packages/explorer/src/components/ui/Form.tsx
@@ -5,7 +5,7 @@ import { Controller, ControllerProps, FieldPath, FieldValues, FormProvider, useF
 import * as LabelPrimitive from "@radix-ui/react-label";
 import { Slot } from "@radix-ui/react-slot";
 import { Label } from "../../components/ui/Label";
-import { cn } from "../../lib/utils";
+import { cn } from "../../utils";
 
 const Form = FormProvider;