+++ packages/block-logs-stream/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @latticexyz/block-logs-stream
 
+## 2.2.10
+
+### Patch Changes
+
+- @latticexyz/common@2.2.10
+
+## 2.2.9
+
+### Patch Changes
+
+- @latticexyz/common@2.2.9
+
 ## 2.2.8
 
 ### Patch Changes