+++ packages/store/CHANGELOG.md
@@ -1,5 +1,23 @@
 # Change Log
 
+## 2.2.10
+
+### Patch Changes
+
+- @latticexyz/common@2.2.10
+- @latticexyz/config@2.2.10
+- @latticexyz/protocol-parser@2.2.10
+- @latticexyz/schema-type@2.2.10
+
+## 2.2.9
+
+### Patch Changes
+
+- @latticexyz/common@2.2.9
+- @latticexyz/config@2.2.9
+- @latticexyz/protocol-parser@2.2.9
+- @latticexyz/schema-type@2.2.9
+
 ## 2.2.8
 
 ### Patch Changes